mas_2/requirements.txt
langgraph
langchain-openai
langchain
langchain-core
python-dotenv
pydantic
chromadb
sentence-transformers
requests
gseapy
pandas
pytest
pytest-mock

